requirements.txt
Flask
redis
azure-cosmos
numpy
scikit-learn
